backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9  # PostgreSQL adapter
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
cryptography>=41.0.0  # For Circle API entity secret encryption
apscheduler>=3.10.4  # For scheduled payroll tasks
# alembic>=1.12.1  # Optional, for database migrations

